fix: tx timeout may fail when time are not synchronized among metanodes

+++ proto/admin_proto.go
@@ -800,7 +800,7 @@ type SimpleVolView struct {
 	EnableToken           bool
 	EnablePosixAcl        bool
 	EnableTransaction     string
-	TxTimeout             uint32
+	TxTimeout             int64
 	Description           string
 	DpSelectorName        string
 	DpSelectorParm        string
